Rename constants module as config
Config reflects better the purpose of the module.
