+++ ci/infra/testrunner/utils/__init__.py
@@ -1,4 +1,4 @@
-from utils.constants import (BaseConfig, Constant)
+from utils.config import (BaseConfig, Constant)
 from utils.format import Format
 from utils.logger import Logger
 from utils.utils import (Utils, step)
