+++ ci/infra/testrunner/utils/utils.py
@@ -11,7 +11,7 @@ from threading import Thread
 import requests
 from timeout_decorator import timeout
 
-from utils.constants import Constant
+from utils.config import Constant
 from utils.format import Format
 
 logger = logging.getLogger('testrunner')
